ARMCI fortran sources use integer*4 rather than integer compiler flag

+++ armci/src/common/fcopy.f
@@ -1,6 +1,6 @@
       subroutine dcopy21(rows, cols, A, ald, buf, cur)
-      integer rows, cols
-      integer c, r, ald, cur 
+      integer*4 rows, cols
+      integer*4 c, r, ald, cur 
       double precision A(ald,*), buf(ald) 
       cur = 0
       do c = 1, cols
@@ -12,8 +12,8 @@
       end
 
       subroutine dcopy31(rows, cols, planes, A, aldr, aldc, buf, cur)
-      integer rows, cols, planes
-      integer c, r, p, aldr, aldc, cur
+      integer*4 rows, cols, planes
+      integer*4 c, r, p, aldr, aldc, cur
       double precision A(aldr, aldc, *), buf(aldr)
       cur = 0
       do p = 1, planes 
@@ -27,8 +27,8 @@
       end
 
       subroutine dcopy12(rows, cols, A, ald, buf, cur)
-      integer rows, cols
-      integer c, r, ald, cur
+      integer*4 rows, cols
+      integer*4 c, r, ald, cur
       double precision A(ald,*), buf(ald)
       cur = 0
       do c = 1, cols
@@ -40,8 +40,8 @@
       end
 
       subroutine dcopy13(rows, cols, planes, A, aldr, aldc, buf, cur)
-      integer rows, cols, planes
-      integer c, r, p, aldr, aldc, cur
+      integer*4 rows, cols, planes
+      integer*4 c, r, p, aldr, aldc, cur
       double precision A(aldr, aldc, *), buf(aldr)
       cur = 0
       do p = 1, planes
@@ -55,8 +55,8 @@
       end
 
       subroutine dcopy2d_n(rows, cols, A, ald, B, bld)
-      integer rows, cols
-      integer c, r, ald, bld
+      integer*4 rows, cols
+      integer*4 c, r, ald, bld
       double precision A(ald,*), B(bld,*)
       do c = 1, cols
          do r = 1, rows
@@ -66,10 +66,10 @@
       end
 
       subroutine dcopy2d_u(rows, cols, A, ald, B, bld)
-      integer rows, cols
-      integer c, r, ald, bld
+      integer*4 rows, cols
+      integer*4 c, r, ald, bld
       double precision A(ald,*), B(bld,*)
-      integer r1
+      integer*4 r1
       double precision d1, d2, d3, d4
       do c = 1, cols
       r1 = iand(max0(rows,0),3)
@@ -95,7 +95,7 @@ c$$$         b(r+3,c) = a(r+3,c) + b(r+3,c) * 0
       end
 
       subroutine dcopy1d_n(A, B, n)
-      integer n,i 
+      integer*4 n,i 
       double precision A(n), B(n)
 ccdir$ no_cache_alloc a,b
       do i = 1, n 
@@ -104,7 +104,7 @@ ccdir$ no_cache_alloc a,b
       end
 
       subroutine dcopy1d_u(A, B, n)
-      integer n,n1,i
+      integer*4 n,n1,i
       double precision A(n), B(n)
       double precision d1, d2, d3, d4
       n1 = iand(max0(n,0),3)
